add "-Wall -Wextra -Werror"

+++ mininn/kernel/add_compute.cc
@@ -33,7 +33,7 @@ void add_func_multi_threads(float* x_buffer, float* y_buffer, float* out_buffer,
     std::vector<std::thread> threads;
     int chunk_size = size / num_threads;
 
-    for (int i = 0; i < num_threads; ++i) {
+    for (size_t i = 0; i < num_threads; ++i) {
         int start = i * chunk_size;
         int end = (i == num_threads - 1) ? size : start + chunk_size;
         threads.emplace_back(add_func, x_buffer, y_buffer, out_buffer, start, end);
